# auto/api/v1/pub.go
// Code generated by go-mir. DO NOT EDIT.
// versions:
// - mir 5.3

package v1

import (
	"net/http"

	"github.com/alimy/mir/v5"
	"github.com/gin-gonic/gin"
	"github.com/rocboss/paopao-ce/internal/model/web"
)

type Pub interface {
	_default_

	SendCaptcha(*web.SendCaptchaReq) error
	GetCaptcha() (*web.GetCaptchaResp, error)
	Register(*web.RegisterReq) (*web.RegisterResp, error)
	Login(*web.LoginReq) (*web.LoginResp, error)
	Version() (*web.VersionResp, error)

	mustEmbedUnimplementedPubServant()
}

// RegisterPubServant register Pub servant to gin
func RegisterPubServant(e *gin.Engine, s Pub) {
	router := e.Group("v1")

	// register routes info to router
	router.Handle("POST", "/captcha", func(c *gin.Context) {
		select {
		case <-c.Request.Context().Done():
			return
		default:
		}
		req := new(web.SendCaptchaReq)
		if err := s.Bind(c, req); err != nil {
			s.Render(c, nil, err)
			return
		}
		s.Render(c, nil, s.SendCaptcha(req))
	})
	router.Handle("GET", "/captcha", func(c *gin.Context) {
		select {
		case <-c.Request.Context().Done():
			return
		default:
		}

		resp, err := s.GetCaptcha()
		s.Render(c, resp, err)
	})
	router.Handle("POST", "/auth/register", func(c *gin.Context) {
		select {
		case <-c.Request.Context().Done():
			return
		default:
		}
		req := new(web.RegisterReq)
		if err := s.Bind(c, req); err != nil {
			s.Render(c, nil, err)
			return
		}
		resp, err := s.Register(req)
		s.Render(c, resp, err)
	})
	router.Handle("POST", "/auth/login", func(c *gin.Context) {
		select {
		case <-c.Request.Context().Done():
			return
		default:
		}
		req := new(web.LoginReq)
		if err := s.Bind(c, req); err != nil {
			s.Render(c, nil, err)
			return
		}
		resp, err := s.Login(req)
		s.Render(c, resp, err)
	})
	router.Handle("GET", "/", func(c *gin.Context) {
		select {
		case <-c.Request.Context().Done():
			return
		default:
		}

		resp, err := s.Version()
		s.Render(c, resp, err)
	})
}

// UnimplementedPubServant can be embedded to have forward compatible implementations.
type UnimplementedPubServant struct{}

func (UnimplementedPubServant) SendCaptcha(req *web.SendCaptchaReq) error {
	return mir.Errorln(http.StatusNotImplemented, http.StatusText(http.StatusNotImplemented))
}

func (UnimplementedPubServant) GetCaptcha() (*web.GetCaptchaResp, error) {
	return nil, mir.Errorln(http.StatusNotImplemented, http.StatusText(http.StatusNotImplemented))
}

func (UnimplementedPubServant) Register(req *web.RegisterReq) (*web.RegisterResp, error) {
	return nil, mir.Errorln(http.StatusNotImplemented, http.StatusText(http.StatusNotImplemented))
}

func (UnimplementedPubServant) Login(req *web.LoginReq) (*web.LoginResp, error) {
	return nil, mir.Errorln(http.StatusNotImplemented, http.StatusText(http.StatusNotImplemented))
}

func (UnimplementedPubServant) Version() (*web.VersionResp, error) {
	return nil, mir.Errorln(http.StatusNotImplemented, http.StatusText(http.StatusNotImplemented))
}

func (UnimplementedPubServant) mustEmbedUnimplementedPubServant() {}