# docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "paopao-ce API",
    "description": "# paopao-ce API 文档\n\n欢迎使用 paopao-ce API 文档！\n\npaopao-ce 是一个开源的泡泡社区系统，提供完整的社交媒体功能。\n\n## 鉴权说明\n\n大部分API需要JWT Token进行鉴权，请在请求头中包含：\n```\nAuthorization: Bearer <your-jwt-token>\n```\n\n## 响应格式\n\n所有API响应都遵循以下格式：\n```json\n{\n  \"code\": 0,\n  \"message\": \"success\",\n  \"data\": {}\n}\n```\n\n## 错误码说明\n\n- `0`: 成功\n- `10001`: 未授权\n- `10002`: 参数错误\n- `10003`: 权限不足\n- `10004`: 资源不存在\n\n## 分页参数\n\n支持分页的API使用以下查询参数：\n- `page`: 页码 (默认: 1)\n- `page_size`: 每页数量 (默认: 20)\n",
    "version": "1.0.0",
    "contact": {
      "name": "paopao-ce",
      "url": "https://github.com/rocboss/paopao-ce"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8008",
      "description": "本地开发环境"
    },
    {
      "url": "https://api.paopao.info",
      "description": "生产环境"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["公开接口"],
        "summary": "用户登录",
        "description": "用户账号密码登录，无需鉴权\n\n**注意事项：**\n- 用户名和密码都是必填项\n- 登录成功后会返回JWT Token",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResp"
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "用户名或密码错误"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["公开接口"],
        "summary": "用户注册",
        "description": "用户账号注册，无需鉴权\n\n**注意事项：**\n- 用户名和密码都是必填项\n- 用户名不能重复",
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResp"
                }
              }
            }
          },
          "400": {
            "description": "参数错误或用户名已存在"
          }
        }
      }
    },
    "/captcha": {
      "get": {
        "tags": ["公开接口"],
        "summary": "获取图片验证码",
        "description": "获取图片验证码用于后续验证，无需鉴权\n\n**响应说明：**\n- `id`: 验证码ID，用于后续验证\n- `b64s`: Base64编码的验证码图片",
        "operationId": "getCaptcha",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCaptchaResp"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["公开接口"],
        "summary": "发送短信验证码",
        "description": "发送短信验证码到指定手机号，无需鉴权\n\n**注意事项：**\n- 需要先获取图片验证码并验证通过\n- 手机号格式必须正确",
        "operationId": "sendCaptcha",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendCaptchaReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "验证码发送成功"
          },
          "400": {
            "description": "参数错误或图片验证码验证失败"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "GetCaptchaResp": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "验证码ID"
          },
          "b64s": {
            "type": "string",
            "description": "Base64编码的验证码图片"
          }
        }
      },
      "SendCaptchaReq": {
        "type": "object",
        "required": ["phone", "img_captcha", "img_captcha_id"],
        "properties": {
          "phone": {
            "type": "string",
            "description": "手机号"
          },
          "img_captcha": {
            "type": "string",
            "description": "图片验证码"
          },
          "img_captcha_id": {
            "type": "string",
            "description": "图片验证码ID"
          }
        }
      },
      "LoginReq": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "description": "密码"
          }
        }
      },
      "LoginResp": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT Token"
          }
        }
      },
      "RegisterReq": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "description": "密码"
          }
        }
      },
      "RegisterResp": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "用户ID"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          }
        }
      },
...
    }
  }
}